pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "robust_metrics"
authors = [
    {name = "Ahmed Chabib", email = "ahmed.chabib16@gmail.com"},
]
maintainers = [
    {name = "Antoine Weisrock", email = "antoine.weisrock@gmail.com"},
]
description ="A python module for robust optical flow measurement."
readme = "README.md"
requires-python = ">=3.6"
keywords = ["Optical Flow", "Correlation", "GPU"]
license-files = ["LICENSE.txt"]
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy",
    "cupy-cuda12x",
    "scipy",
    "opencv-python",
    "matplotlib",
    "cucim"
]

version = "0.0.1"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false
